+++ src/backend/marsha/core/xapi.py
@@ -71,8 +71,12 @@ class XAPIStatement:
         }
 
         statement["object"] = {
-            "definition": {"type": "https://w3id.org/xapi/video/activity-type/video"},
-            "name": {to_locale(settings.LANGUAGE_CODE).replace("_", "-"): video.title},
+            "definition": {
+                "type": "https://w3id.org/xapi/video/activity-type/video",
+                "name": {
+                    to_locale(settings.LANGUAGE_CODE).replace("_", "-"): video.title
+                },
+            },
             "id": "uuid://{id}".format(id=str(video.id)),
             "objectType": "Activity",
         }